+++ projects/rxjs-supersets/src/delta-map/delta-map.spec.ts
@@ -1,6 +1,6 @@
 import { Subscription } from 'rxjs';
 
-import { processDelta } from '../process-delta/process-delta';
+import { processDelta } from '../operators/process-delta/process-delta';
 import { IdObject, MapDelta } from '../types';
 
 import { DeltaMap } from './delta-map';