+++ projects/rxjs-supersets/src/super-set/super-set.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Subscription } from 'rxjs';
 
-import { processDelta } from '../process-delta/process-delta';
+import { processDelta } from '../operators/process-delta/process-delta';
 import { MemberObject, DeltaObservable, MapDelta } from '../types';
 
 import { SuperSet } from './super-set';